Added in the ability to pass in a ClusterEnvironment
This will allow a user to configure the NRI ES integration to have the cluster name report with extra granularity about which enviroment it's in.
By default this is disabled, and only makes a difference when a user passes in the arg.

A example use case here is that a user has two ES clusters with the same name, but running in different enviroments.